+++ app/AppKernel.php
@@ -68,6 +68,7 @@ class AppKernel extends Kernel
             $bundles[] = new Symfony\Bundle\DebugBundle\DebugBundle();
             $bundles[] = new Symfony\Bundle\WebProfilerBundle\WebProfilerBundle();
             $bundles[] = new Hpatoio\DeployBundle\DeployBundle();
+            $bundles[] = new \Symfony\Bundle\MakerBundle\MakerBundle();
 
             if ('dev' === $this->getEnvironment()) {
                 # $bundles[] = new Sensio\Bundle\GeneratorBundle\SensioGeneratorBundle();