+++ src/Cocorico/CoreBundle/Repository/DirectoryRepository.php
@@ -26,7 +26,8 @@ class DirectoryRepository extends EntityRepository
         $qB->setMaxResults($limit)
            ->setFirstResult($offset)
            ->orderBy('d.name', 'asc')
-           ->where('d.isActive = true');
+           ->where('d.nature != \'n/a\'')
+           ->where('d.isDelisted = false');
 
         return $qB;
     }
@@ -35,7 +36,16 @@ class DirectoryRepository extends EntityRepository
     {
         $qB = $this->getFindQueryBuilder();
         $qB->orderBy('d.name', 'asc')
-           ->where('d.isActive = true');
+           ->where('d.nature != \'n/a\'')
+           ->where('d.isDelisted = false');
+
+        return $qB;
+    }
+
+    public function getFull()
+    {
+        $qB = $this->getFindQueryBuilder();
+        $qB->orderBy('d.name', 'asc');
 
         return $qB;
     }