+++ src/Cocorico/MessageBundle/Entity/Thread.php
@@ -146,4 +146,40 @@ class Thread extends BaseThread
     }
 
 
+
+    /**
+     * Add metadata.
+     *
+     * @param \Cocorico\MessageBundle\Entity\ThreadMetadata $metadata
+     *
+     * @return Thread
+     */
+    public function addMetadatum(\Cocorico\MessageBundle\Entity\ThreadMetadata $metadata)
+    {
+        $this->metadata[] = $metadata;
+
+        return $this;
+    }
+
+    /**
+     * Remove metadata.
+     *
+     * @param \Cocorico\MessageBundle\Entity\ThreadMetadata $metadata
+     *
+     * @return boolean TRUE if this collection contained the specified element, FALSE otherwise.
+     */
+    public function removeMetadatum(\Cocorico\MessageBundle\Entity\ThreadMetadata $metadata)
+    {
+        return $this->metadata->removeElement($metadata);
+    }
+
+    /**
+     * Get metadata.
+     *
+     * @return \Doctrine\Common\Collections\Collection
+     */
+    public function getMetadata()
+    {
+        return $this->metadata;
+    }
 }